Apply large changes to the stucture of HyperBinningHistograms (now known as a HyperHistogram).
Previously a HyperBinningHistogram would have a HyperVolumeBinning as a private member. Now, the
renamed HyperHistogram has a pointer to a BinningBase, so that it can use any binning type. This will
allow the creation of a DiskResident binning that can be used instead of a HyperVolumeBinning. In
addition the HyperVolumeBinning is now renamed as HyperBinning for simplicity.

	modified:   example/src/TestBinning.cpp

Updated example to work with edited code.

	new file:   include/BinningBase.h

New base class for all binnings.

	renamed:    include/HyperVolumeBinning.h -> include/HyperBinning.h

Renamed HyperVolumeBinning as HyperBinning, which now inherents from BinningBase (so can be used with a HyperHistogram)

	modified:   include/HyperBinningMaker.h

Update to take HyperBinning instead of HyperVolumeBinning

	modified:   include/HyperBinningPainter.h

Update to take HyperBinning instead of HyperVolumeBinning

	modified:   include/HyperBinningPainter1D.h

Update to take HyperBinning instead of HyperVolumeBinning

	modified:   include/HyperBinningPainter2D.h

Update to take HyperBinning instead of HyperVolumeBinning

	renamed:    include/HyperBinningHistogram.h -> include/HyperHistogram.h

Renamed, and now store a pointer to a BinningBase rather than a HyperVolumeBinning (not pointer). This allowes a HyperHistogram to use different binning types.